Created View
I have create a view for the map and allowed for the user to move around the window using the WASD keys. I have made a simple box and coloured it in green which will be the baseline for the game's map.

Uisng a new component called GameObject I will be able to create new objects with the ability to be drawn and moved within the game wor;ld easily.

Next I will need to add a grid overtop of the game map (the green box) so that I can move onto the random generation of the map.