Added Grid into game.
I have created two new classes a grid and a cells class each will allow for a grid of varing size to be created onto a rectangle shape within the game world.

Next I need to add in functionality of tiles, allowing for the cells to be given properties, for example "Dirt" or "Water" as the first steps to creating a random map.