Added a class and Placed lake randomly
I have created a new component class for the game, which will be used to generate random numbers easily. To follow this I have also added a segmet of code to generate a seed for the program based on the time.

I have also imoplemented some simple code to randomly select a cell on the grid and pass that into the lake generation code as a start point.

Next I will work on river generation for the game, so at a later date either a river or a lake will be selected for the game. Generating a river will require finding two points, a start and end at the edge of the map and finding a path between them creating a river on the map.