A* pathfindinh algorithm
I have added a working version of A* into the game, it will find the shortest path between two points.

I have currently got this in the map code and will need to do a heavy session of commenting.

I will also need to output the generated path and create a river in the game using it.