Random Generation Step 1
I have completed random generation for a single layer on the map. This includes water, rock and dirt generation for a ground level layer. The core for the generation is this layer, and next I will work on generating a map for multiple different layers. Using what was generated on the ground layer, I can then generate lower and upper layers for the game.

Todo work on generating the lower layers for the map. -1 and -2. (they should be mostly rock with a little water and dirt).