Begun Colonist Work
I have created a placeholder for what will become the game's colonists. I have used a single white box just outside the current grid for the game's map.

Next I will need to randomly select a location on the game map which is a dirt tile to now place the colonist.

To do this I will need to add a variable within a cell for it's center. I may also need to create a vector on grid which keeps track of dirt cells on a given layer.