Colonist Placement
I have finished the colonist placement for the game; upon the game start the colonist will be randomly placed somewhere on the map, within a dirt tile.

I have also ensured that the colonist cannot be placed within a rock tile.

Next I will need to work on integrating A* into the colonists, this will allow for them to move around the map. For this to work I will also need to make a few additions into the A* functionality, adding in the rock tiles as obstructions.