Colonist Path finding
I have added a basic working version of the pathfinding code into the colonist class, this will allow for them to move around the map freely.

Next I will need to to some serious cleanup for the game. A lot of the colonist class remains mostly uncommented and needs it sorting soon.

I also need to perform a round of testing with the pathfinding, see how it works if there is no path to be found, or if i add multiple colonists with different paths as well as the same path.

Good job for today!