Round of commenting
I have commented the main bulk of the new code, as well as the class Colonists.

I need to add in some try catches for the code which may run into null pointers, this will help for the general efficiveness for the code.

I need to implement the colonist movement code once the path has been found, possibly find a line along the path and iterate slowly along that line, updating the colonist's position.

Testing is also a key stage that needs to occur at some point soon, try to break something while in the game; find some bugs to fix!