Try Catch
I have added some extra levels of security to the code by the form of try catch blocks whenever a pointer can be passed into a function. These blocks will check to see if the pointer is a nullptr and pass over the function is it is; preventing any errors which may occur.

I have added a new folder which will be used to sore the game's documentation. This will allow for a easy split between the game code and the game's other douments. I have also made a basic notepad document which will store the error codes; at the moment the error codes have a basic line saying what it is for but at a later date I will explain these codes in greater detail.

Next, I will work on moving the colonist along the generated path.