Colonist Movement
I have enabled the colonists to use A* and move around the map, they currently continuously move around the map generating a new path once they reach the end of their current one. I have also added in a timeout value for loops within the game, so if a loop remains active for a prolonged period of time it will timeout and be forced to end.

Next I will need to make some changes to the A* code (possibly create an overload) to include obsticales to the game, this will prevent the colonists from walking into 'solid' objects like likes with rock on them.

I also need to do a heavy round of commenting within the functions to make it far easier for me to know what is going on.