Obstructions
I have added a list of obstructions which can be passed into the algorithm and any tiles maked with these will be avoided by the algorithm.

Next I will need to make some improvements to the algorithm operates, because as more colonists are added it becomes quite unplayable because the game pauses while a path is found. I need to make a better way for the path finding process to occur without the loss of accuracy of many colonist's using the same map.