Threading
I have added a new thread for the pathfinding functionality. This will allow for the pathfinding algorithm to run indipendently from the main game loop, this will allow for the whole path to be generated without pausing the game loop.

Next, I will work on making a job system into the game, hopefully this will help the expansion of the colonists later. I will start off with an idle job which will make them move around the map randomly.

I will also need to conduct a little more research into the threading, to ensure that my implementation is correct.