Slight Chnges
I have removed the timeout console print, and assigned some constraints to the colonists making them only be drawn on the layer they are currently on.

Next, I think i'll add a line from the colonists current possition to the end of their path, making it clear where they are going without changing the colour of the map's tiles.