Cleanup
I have done a little cleanup and some colour review. I found that the sky and the water were too similar to each other and edited the colour of the water.

Next, I aim to add in a line from the colonist's current position to their endpoint making it easier to see where they are going.