Added Line
I have added a line strip for the colonists, this will now show the player where the colonist will be moving to and the path they will take to get there.

Following this, I will want to add a job system into the game, so I can convert this random wandering into an Idle job so that new jobs can easily be added to allow for the colonists to perform different functions at different times.

:+1: