Job System
I have added the base requirements for a job system into the game and converted the current code into that system. The only current job in the game is an idle job with no actions fitted into thta job.

Next, I plan to add in some resource generation into the game, starting with trees randomly planted on the dirt within the ground level of the map.