Finding nearest tree
I am using a literal distance equation to find the clostest tree in the game world and so that means that teh time it takes to get to a tree isn't taken into account; this has the possibility of the colonist traveling through a potentially farther away tree in order to get to the target. with some time I could refine this further but it works how I need it at the moment.

Next the process of cutting the tree down and replacing it with some resources within the game world.