Cutting trees
I have allowed for the colonist to cut down trees within the game world; the only issue is that when the resource is dropped onto the floor the text with the number is not displayed and thus breaks the game.

This issue needs to be addressed as soon as possible.