Build and Wood Pile Text
I have fixed the issue wherein when the game attempts to draw the text on the wood piles it would break; I have also adjusted the pathfinding system to when the colonist looks for the tree if it can't find one after a few tries they go back to being idle.

I have also created build version 0.01 for basic external testing purposes.