Initiated the Building objects
I have created two new classes, one which will be an individual building object, like a wall and a second with the aims to hold all of the buldings in a single place.

I have begun with the basics drawing and the such now I need to add in a system to allow for the manager to create and remove building within the game.