Simple building system
I have made it so that the player can place squares into the game world (Placeholder buildings), this will form the basis for the building process; next I need to have the colonist move over to the building and build it. I aim to use similar functionality to the tree cutting to make them move.

Following I will need to make a system to potentially add the new walls into a list of obstructions (or just the specific cells) and prevent the colonist from moving through them.