Building Objects
I have allowed the player to assign buildings to be constructed and the colonist will now travel over to them and build them. I need to test to ensure the colonist will not move through the walls though.

I will also need to prevent assigning buildings when clicking on buttons (maybe if the mouse is within a certain y, something to mess with).

Next texturing the buildings is important to seperate the different building types.