Added a sleep counter
I have added in a value which will be used to count the colonist's need for sleep, thsi will slowly go down through the course of the game and when it reaches 0 the colonist will begin to die. I will need to add in a bed object to allow for the colonist to replenish their sleep meter.